requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0

# AI/ML dependencies
openai>=1.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.8.0
numpy>=1.21.0

# HTTP client
httpx>=0.24.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Optional: For better JSON handling
orjson>=3.8.0

# Optional: For environment management
python-dotenv>=1.0.0
